requirements.txt
Django==1.9
Pillow==3.3.0
amqp==1.4.9
celery==3.1.23
django-autoslug==1.9.3
django-celery==3.1.17
django-grappelli==2.8.1
feedparser==5.2.1
mysqlclient==1.3.7
requests==2.10.0
anyjson==0.3.3
dateutils==0.6.6
django-social-widgets==0.4.0
django-endless-pagination
djangorestframework